moving name config to an additional file
